fix #1
The function file_is_newer used to compare mtime of the manifest to the
mtime of files. mtime is not updated when files change permissions or
names, so the comparison was changed to ctime.

Also fixes a bug with any files containing .md rendering as Markdown
files, even if they are not actually .md files or if they are backups.